web-app/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
opencv-python>=4.8.0 
mediapipe
numpy
matplotlib
transformers
Pillow
tqdm
supervision
google-generativeai